+++ cli/azd/pkg/config/file_config_manager_test.go
@@ -97,6 +97,19 @@ func Test_FileConfigManager_GetSetSecrets(t *testing.T) {
 	missingPassword, ok := updatedConfig.GetString("secrets.misingVaultRef")
 	require.False(t, ok)
 	require.Empty(t, missingPassword)
+
+	require.NotEqual(t, updatedConfig.Raw(), updatedConfig.ResolvedRaw())
+	require.Equal(t, azdConfig.Raw(), updatedConfig.Raw())
+	require.Equal(t, azdConfig.ResolvedRaw(), updatedConfig.ResolvedRaw())
+
+	// verify vault reference
+	vault, exists := azdConfig.GetString(vaultKeyName)
+	require.True(t, exists)
+	require.NotEmpty(t, vault)
+	configFromRaw := NewConfig(azdConfig.ResolvedRaw())
+	vault, exists = configFromRaw.GetString(vaultKeyName)
+	require.False(t, exists)
+	require.Empty(t, vault)
 }
 
 func Test_FileConfigManager_GetSetSecretsInSection(t *testing.T) {