+++ cli/azd/pkg/project/framework_service_docker.go
@@ -371,6 +371,10 @@ func (p *dockerProject) packBuild(
 		// Always default to port 80 for consistency across languages
 		environ = append(environ, "ORYX_RUNTIME_PORT=80")
 
+		if svc.Language == ServiceLanguageJava {
+			environ = append(environ, "ORYX_RUNTIME_PORT=8080")
+		}
+
 		if svc.OutputPath != "" && (svc.Language == ServiceLanguageTypeScript || svc.Language == ServiceLanguageJavaScript) {
 			inDockerOutputPath := path.Join("/workspace", svc.OutputPath)
 			// A dist folder has been set.