+++ cli/azd/pkg/project/service_target_ai_endpoint.go
@@ -161,15 +161,26 @@ func (m *aiEndpointTarget) Deploy(
 				return
 			}
 
+			if onlineDeployment == nil {
+				task.SetError(fmt.Errorf("unexpected response from deployToEndpoint: deployment is nil"))
+				return
+			}
+			if onlineDeployment.Name == nil {
+				task.SetError(fmt.Errorf("unexpected response from deployToEndpoint: deployment name is nil"))
+				return
+			}
+
+			deploymentName := *onlineDeployment.Name
 			task.SetProgress(NewServiceProgress("Updating traffic"))
-			_, err = m.aiHelper.UpdateTraffic(ctx, workspaceScope, endpointName, *onlineDeployment.Name)
+			_, err = m.aiHelper.UpdateTraffic(ctx, workspaceScope, endpointName, deploymentName)
 			if err != nil {
 				task.SetError(fmt.Errorf("failed updating traffic: %w", err))
 				return
 			}
 
 			task.SetProgress(NewServiceProgress("Removing old deployments"))
-			if err := m.aiHelper.DeletePreviousDeployments(ctx, workspaceScope, endpointName); err != nil {
+			if err := m.aiHelper.DeleteDeployments(
+				ctx, workspaceScope, endpointName, []string{deploymentName}); err != nil {
 				task.SetError(fmt.Errorf("failed deleting previous deployments: %w", err))
 				return
 			}
@@ -206,9 +217,28 @@ func (m *aiEndpointTarget) Endpoints(
 		return nil, fmt.Errorf("failed initializing AI project: %w", err)
 	}
 
+	tenantId, has := m.env.LookupEnv(environment.TenantIdEnvVarName)
+	if !has {
+		return nil, fmt.Errorf(
+			"tenant ID not found. Ensure %s has been set in the environment.",
+			environment.TenantIdEnvVarName,
+		)
+	}
+
 	workspaceScope, err := m.getWorkspaceScope(serviceConfig, targetResource)
 	if err != nil {
 		return nil, err
+	}
+
+	workspaceLink := ai.AiStudioWorkspaceLink(
+		tenantId,
+		workspaceScope.SubscriptionId(),
+		workspaceScope.ResourceGroup(),
+		workspaceScope.Workspace(),
+	)
+
+	endpoints := []string{
+		fmt.Sprintf("Workspace: %s", workspaceLink),
 	}
 
 	endpointName := filepath.Base(targetResource.ResourceName())
@@ -217,10 +247,34 @@ func (m *aiEndpointTarget) Endpoints(
 		return nil, err
 	}
 
-	return []string{
+	var deploymentName string
+	for key, value := range onlineEndpoint.Properties.Traffic {
+		if *value == 100 {
+			deploymentName = key
+			break
+		}
+	}
+
+	if deploymentName != "" {
+		deploymentLink := ai.AiStudioDeploymentLink(
+			tenantId,
+			workspaceScope.SubscriptionId(),
+			workspaceScope.ResourceGroup(),
+			workspaceScope.Workspace(),
+			endpointName,
+			deploymentName,
+		)
+
+		endpoints = append(endpoints, fmt.Sprintf("Deployment: %s", deploymentLink))
+	}
+
+	endpoints = append(
+		endpoints,
 		fmt.Sprintf("Scoring: %s", *onlineEndpoint.Properties.ScoringURI),
 		fmt.Sprintf("Swagger: %s", *onlineEndpoint.Properties.SwaggerURI),
-	}, nil
+	)
+
+	return endpoints, nil
 }
 
 // getWorkspaceScope returns the scope for the workspace